Split MissingElements into its own public class
